[20] implement basic breadcrumbs
Uses default bootstrap styling.

changed how routes are organized to make dynamic breadcrumbs easy. Did have some trouble at first but all the issues seem solved now... I'd have liked to solve this without the "empty" Home component but the transitions between pages still seem smooth to me so, not too bad?

Be aware of the url/page changing(or not) when trying to go between "/" and "/upload" and vice versa. This is not unexpected (/upload is the desired home page, after all)